+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 
 ## Version 7.2
 
+### Version 7.2.2.0 - 2024-10-31
+
+-   Added support for a fully SQLAlchemy object driven external table creation
+    example
+-   Fixed `TINYINT` type mapping
+-   Added link to SQLAlchemy section of Kinetica documentation site
+-   Fixed boolean parameter handling in example scripts
+-   Removed invalid JSON array type from examples
+
+
 ### Version 7.2.1.1 - 2024-09-18
 
 -   Made PIVOT & UNPIVOT processors supportive of more input table/query objects